requirements.txt
Pillow==2.3.0
nose==1.3.1
six==1.6.1
Sphinx==1.2.2

